Изменение типов моделей
HEALTHDIARY-55: MetricAnalysis

+++ HealthDiary/MetricService.BLL/DTO/AnalysisResult/AnalysisResultBaseDTO.cs
@@ -11,7 +11,7 @@
         /// <summary>
         /// Числовое значение результата анализа
         /// </summary>        
-        public decimal? Value { get; set; }
+        public float? Value { get; set; }
 
 
         /// <summary>
